Move declarative auth mapping into plugins service
New interface:
- declarative.MappedCredentialParser(mapping) maps structured credential JSON into an auth header and extra headers for declarative plugin execution.

Example:
base.TokenParser = declarative.MappedCredentialParser(&providermanifestv1.AuthMapping{
    Headers: map[string]providermanifestv1.AuthValue{
        X-Api-Key: {ValueFrom: &providermanifestv1.AuthValueFrom{CredentialFieldRef: &providermanifestv1.CredentialFieldRef{Name: api_key}}},
    },
})

The legacy internal provider builder now consumes the service-owned parser, and the manifest-backed declarative provider no longer imports internal/provider.
